requirements.txt
tensorflow
numpy
opencv-python
uvicorn
fastapi
python-multipart
requests
pillow
pydantic